requirements.txt
# Runtime dependencies for app.py, ingest.py, rag.py (Streamlit Community Cloud + local)
streamlit>=1.28.0
# sentence-transformers -> huggingface transformers may import vision helpers; Streamlit's watcher
# can trigger those code paths → ModuleNotFoundError: torchvision without explicit installs.
torch
torchvision
sentence-transformers
chromadb
pypdf
python-dotenv
google-genai
